fix spark rapids ml and spark dependencies
