+++ python/benchmark/benchmark/bench_nearest_neighbors.py
@@ -24,59 +24,6 @@ from benchmark.base import BenchmarkBase
 from benchmark.utils import with_benchmark
 
 
-# create a dummy class to avoid spark_rapids_ml import at top level for
-# spark connect client environments
-class NearestNeighborsBenchmarkClasses:
-    from spark_rapids_ml.knn import ApproximateNearestNeighborsModel
-
-    class CPUNearestNeighborsModel(ApproximateNearestNeighborsModel):
-        from spark_rapids_ml.core import (
-            EvalMetricInfo,
-            _ConstructFunc,
-            _EvaluateFunc,
-            _TransformFunc,
-            alias,
-        )
-
-        def __init__(self, item_df: DataFrame):
-            super().__init__(item_df)
-
-        def kneighbors(
-            self, query_df: DataFrame, sort_knn_df_by_query_id: bool = True
-        ) -> Tuple[DataFrame, DataFrame, DataFrame]:
-            self._item_df_withid = self._ensureIdCol(self._item_df_withid)
-            return super().kneighbors(
-                query_df, sort_knn_df_by_query_id=sort_knn_df_by_query_id
-            )
-
-        def _get_cuml_transform_func(
-            self,
-            dataset: DataFrame,
-            eval_metric_info: Optional[EvalMetricInfo] = None,
-        ) -> Tuple[
-            _ConstructFunc,
-            _TransformFunc,
-            Optional[_EvaluateFunc],
-        ]:
-            self._cuml_params["algorithm"] = "brute"
-            _, _transform_internal, _ = super()._get_cuml_transform_func(
-                dataset, eval_metric_info
-            )
-
-            from sklearn.neighbors import NearestNeighbors as SKNN
-
-            n_neighbors = self.getK()
-
-            def _construct_sknn() -> SKNN:
-                nn_object = SKNN(algorithm="brute", n_neighbors=n_neighbors)
-                return nn_object
-
-            return _construct_sknn, _transform_internal, None
-
-        def _concate_pdf_batches(self) -> bool:
-            return False
-
-
 class BenchmarkNearestNeighbors(BenchmarkBase):
     def _supported_class_params(self) -> Dict[str, Any]:
         params = {"n_neighbors": 200}
@@ -179,20 +126,17 @@ class BenchmarkNearestNeighbors(BenchmarkBase):
             print(f"gpu total took: {total_time} sec")
 
         if num_cpus > 0:
-            assert num_gpus <= 0
+            from .utils_knn import CPUNearestNeighborsModel
 
+            assert num_gpus <= 0
             if not no_cache:
                 (train_df, query_df), prepare_time = with_benchmark(
                     "prepare dataset", lambda: cache_df(train_df, query_df)
                 )
 
-            def get_cpu_model() -> (
-                NearestNeighborsBenchmarkClasses.CPUNearestNeighborsModel
-            ):
-                cpu_estimator = (
-                    NearestNeighborsBenchmarkClasses.CPUNearestNeighborsModel(
-                        train_df
-                    ).setK(params["n_neighbors"])
+            def get_cpu_model() -> CPUNearestNeighborsModel:
+                cpu_estimator = CPUNearestNeighborsModel(train_df).setK(
+                    params["n_neighbors"]
                 )
 
                 return cpu_estimator
@@ -207,8 +151,7 @@ class BenchmarkNearestNeighbors(BenchmarkBase):
                 cpu_model = cpu_model.setInputCols(input_cols)
 
             def cpu_transform(
-                model: NearestNeighborsBenchmarkClasses.CPUNearestNeighborsModel,
-                df: DataFrame,
+                model: CPUNearestNeighborsModel, df: DataFrame
             ) -> DataFrame:
                 (item_df_withid, query_df_withid, knn_df) = model.kneighbors(df)
                 knn_df.count()
